add to existing textDocumentServiceHandler
